+++ plugins/plugin-svelte/plugin.js
@@ -4,6 +4,11 @@ const fs = require('fs');
 const path = require('path');
 const {createMakeHot} = require('svelte-hmr');
 
+const inlineSourcemap = (code, map) =>
+  code +
+  '\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,' +
+  new Buffer(map.toString()).toString('base64');
+
 module.exports = function plugin(snowpackConfig, pluginOptions = {}) {
   const isDev = snowpackConfig.mode !== 'production';
   const useSourceMaps =
@@ -169,6 +174,11 @@ module.exports = function plugin(snowpackConfig, pluginOptions = {}) {
 
       const compiled = svelte.compile(codeToCompile, finalCompileOptions);
       const {js, css} = compiled;
+
+      if (useSourceMaps) {
+        js.code = inlineSourcemap(js.code, js.map);
+      }
+
       const output = {
         '.js': {
           code: js.code,