+++ plugins/plugin-vue/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 <!-- add changelog entries here - do not delete! -->
 
+## @snowpack/plugin-vue@2.5.0 [2021-05-13]
+
+* 4403595e - Sourcemaps (#3271) <Luke Jackson>
+
 ## @snowpack/plugin-vue@2.4.0 [2021-04-02]
 
 * 5d6ee1f3 - Enable CSS Module support in Vue SSR (#3072) <Drew Powers>